# server/router/confluence-router.go
package router

import (
	"fmt"
	"log"
	"net/http"
	"strings"
	"time"

	"github.com/golang-jwt/jwt/v5"
	"github.com/google/uuid"
	"github.com/gorilla/mux"

	. "github.com/seatsurfing/seatsurfing/server/api"
	. "github.com/seatsurfing/seatsurfing/server/repository"
	. "github.com/seatsurfing/seatsurfing/server/util"
)

type ConfluenceServerClaims struct {
	UserName string `json:"user"`
	UserKey  string `json:"key"`
	jwt.RegisteredClaims
}

type ConfluenceRouter struct {
}

func (router *ConfluenceRouter) SetupRoutes(s *mux.Router) {
	s.HandleFunc("/{orgID}/{jwt}", router.serverLogin).Methods("GET")
}

func (router *ConfluenceRouter) serverLogin(w http.ResponseWriter, r *http.Request) {
	vars := mux.Vars(r)
	org, err := GetOrganizationRepository().GetOne(vars["orgID"])
	if err != nil || org == nil {
		SendTextNotFound(w, "text/plain", router.getOrgNotFoundBody())
		return
	}
	sharedSecret, err := GetSettingsRepository().Get(org.ID, SettingConfluenceServerSharedSecret.Name)
	if err != nil || sharedSecret == "" {
		recordAuthEvent(r, &AuthEvent{OrganizationID: org.ID, Method: AuthMethodConfluence, ErrorCode: AuthErrorIdpConfigInvalid, ErrorDetail: "Confluence shared secret is not configured"})
		SendBadRequest(w)
		return
	}
	claims := &ConfluenceServerClaims{}
	token, err := jwt.ParseWithClaims(vars["jwt"], claims, func(token *jwt.Token) (interface{}, error) {
		if _, ok := token.Method.(*jwt.SigningMethodHMAC); !ok {
			return nil, fmt.Errorf("unexpected signing method: %v", token.Header["alg"])
		}
		return []byte(sharedSecret), nil
	})
	primaryDomain, _ := GetOrganizationRepository().GetPrimaryDomain(org)
	if err != nil {
		log.Println("JWT header verification failed: parsing JWT failed with: " + err.Error())
		recordAuthEvent(r, &AuthEvent{OrganizationID: org.ID, Method: AuthMethodConfluence, ErrorCode: AuthErrorConfluenceJwtInvalid, ErrorDetail: "parsing JWT failed: " + err.Error()})
		SendTemporaryRedirect(w, FormatURL(primaryDomain.DomainName)+"/ui/login/failed/")
		return
	}
	if !token.Valid {
		log.Println("JWT header verification failed: invalid JWT")
		recordAuthEvent(r, &AuthEvent{OrganizationID: org.ID, Method: AuthMethodConfluence, ErrorCode: AuthErrorConfluenceJwtInvalid, ErrorDetail: "invalid JWT"})
		SendTemporaryRedirect(w, FormatURL(primaryDomain.DomainName)+"/ui/login/failed/")
		return
	}
	allowAnonymous, _ := GetSettingsRepository().GetBool(org.ID, SettingConfluenceAnonymous.Name)
	userID := router.getUserEmailServer(org, claims, allowAnonymous)
	if userID == "" {
		SendTemporaryRedirect(w, FormatURL(primaryDomain.DomainName)+"/ui/login/confluence/anonymous")
		return
	}
	_, err = GetUserRepository().GetByAtlassianID(userID)
	if err != nil {
		// user not found using atlassianID, try by mail
		u, err := GetUserRepository().GetByEmail(org.ID, userID)
		if err == nil {
			// got it, update it now
			GetUserRepository().UpdateAtlassianClientIDForUser(u.OrganizationID, u.ID, userID)
		}
		// and load again
		GetUserRepository().GetByAtlassianID(userID)
	}
	if err != nil {
		if !GetUserRepository().CanCreateUser(org) {
			recordAuthEvent(r, &AuthEvent{OrganizationID: org.ID, Email: userID, Method: AuthMethodConfluence, ErrorCode: AuthErrorUserLimitReached})
			SendTemporaryRedirect(w, FormatURL(primaryDomain.DomainName)+"/ui/login/failed")
			return
		}
		user := &User{
			Email:          userID,
			AtlassianID:    NullString(userID),
			OrganizationID: org.ID,
		}
		if err := GetUserRepository().Create(user); err != nil {
			recordAuthEvent(r, &AuthEvent{OrganizationID: org.ID, Email: userID, Method: AuthMethodConfluence, ErrorCode: AuthErrorUserCreateFailed, ErrorDetail: err.Error()})
		}
	}
	payload := &AuthStateLoginPayload{
		LoginType: "",
		UserID:    userID,
	}
	authState := &AuthState{
		AuthProviderID: GetSettingsRepository().GetNullUUID(),
		Expiry:         time.Now().Add(time.Minute * 5),
		AuthStateType:  AuthAtlassian,
		Payload:        marshalAuthStateLoginPayload(payload),
	}
	if err := GetAuthStateRepository().Create(authState); err != nil {
		recordAuthEvent(r, &AuthEvent{OrganizationID: org.ID, Email: userID, Method: AuthMethodConfluence, ErrorCode: AuthErrorInternal, ErrorDetail: "failed to create auth state: " + err.Error()})
		SendInternalServerError(w)
		return
	}
	SendTemporaryRedirect(w, FormatURL(primaryDomain.DomainName)+"/ui/login/success/"+authState.ID+"/")
}

func (router *ConfluenceRouter) getOrgNotFoundBody() []byte {
	var sb strings.Builder
	sb.WriteString("Instance ID could not be found at this instance. ")
	sb.WriteString("\n\n")
	sb.WriteString("To get your Intance ID, log in to the Admin interface, go to Settings and copy the Instance ID.")
	sb.WriteString("\n\n")
	sb.WriteString("Make sure the Instance ID is set under 'Seatsurfing Configuration' in your Confluence installation.")
	sb.WriteString("\n\n")
	sb.WriteString("For more information, please read the documentation at: https://docs.seatsurfing.io/")
	sb.WriteString("\n")
	return []byte(sb.String())
}

func (router *ConfluenceRouter) getUserEmailServer(org *Organization, claims *ConfluenceServerClaims, allowAnonymous bool) string {
	userAccountID := ""
	desiredDomain := ""
	if claims.UserName != "" {
		if ValidateEmail(claims.UserName) {
			mailparts := strings.Split(claims.UserName, "@")
			userAccountID = mailparts[0]
			desiredDomain = mailparts[1]
		}
	}
	if userAccountID == "" {
		if claims.UserName != "" {
			userAccountID = "confluence-" + claims.UserName
		}
		if claims.UserName == "" {
			if !allowAnonymous {
				return ""
			}
			userAccountID = "confluence-anonymous-" + uuid.New().String()
		}
	}
	domains, err := GetOrganizationRepository().GetDomains(org)
	if err != nil {
		return ""
	}
	domain := ""
	otherDomain := ""
	for _, curDomain := range domains {
		if curDomain.Active {
			otherDomain = curDomain.DomainName
			if desiredDomain != "" && desiredDomain == curDomain.DomainName {
				domain = curDomain.DomainName
			}
		}
	}
	if domain == "" {
		domain = otherDomain
	}
	email := userAccountID + "@" + domain
	if !ValidateEmail(email) {
		return ""
	}
	return email
}