# specs/openapi.yaml
openapi: 3.1.0
info:
  title: Seatsurfing REST API
  description: |
    Seatsurfing provides REST APIs used by the mobile and web app as well as the administrator web interface.

    ## Authorization checks

    Each API endpoint is protected by authorization checks:
    1. **Permission check**: the permissions granted by the roles assigned to the calling user
    2. **Organization check**: organization membership - no request reaches data belonging to another organization

    ## Permissions

    Access is granted by roles, which an organization's administrators define. A role grants each
    functionality at a level, and a user may hold several roles, holding the highest level any of
    them grants.

    Levels: `0` no access, `10` read only, `20` limited, `30` full access. Not every level is
    meaningful for every permission; `GET /role/permissions` reports the levels each one offers.

    Permission keys: `areas`, `space_attributes`, `bookings`, `approvals`, `analytics`,
    `presence_report`, `users`, `groups`, `roles`, `service_accounts`, `auth_providers`,
    `org_settings`, `audit_log`. Plugins contribute further keys prefixed with `plugin.`.

    `analytics` covers aggregate occupancy figures; `presence_report` covers who was present
    and when. They are separate because the latter is personal data, and each is additionally
    subject to an organization-wide setting (`hide_stats` and `hide_reports`) that withholds it
...

...

...

    ## Authentication

...

...

...

...

...
  license:
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained via login or refresh endpoints
    BasicAuth:
      type: http
      scheme: basic
      description: Service account authentication using `{orgId}_{email}` as username and the service account password
    ServiceAccountBearerAuth:
      type: http
      scheme: bearer
      description: Service account authentication using a long-lived API token generated via `POST /user/{id}/api-token`. The raw token is presented as a Bearer token. Read-only service accounts may only use this on `GET` requests.
  schemas:
...
    GetDomainResponse:
      type: object
      properties:
        domain:
          type: string
        active:
          type: boolean
        verifyToken:
          type: string
        primary:
          type: boolean
        accessible:
          type: boolean
        accessCheck:
          type:
            - string
            - "null"
          format: date-time
...
  responses:
...
    BadRequest:
      description: Malformed request or incomplete JSON
      headers:
        X-Error-Code:
          schema:
            type: integer
          description: Custom error code (see error codes table)
...
    Forbidden:
      description: Authorized, but user has no access to requested resource
    NotFound:
      description: Object not found
...
paths:
...
  /organization/{id}/domain/:
    get:
      tags: [Organizations]
      summary: Get all domains for an organization
      description: Returns all domains configured for an organization. Requires `org_settings` at full access.
      operationId: getDomains
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of domains
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetDomainResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /confluence/{orgID}/{jwt}:
    get:
      tags: [Confluence]
      summary: Confluence server SSO login
      description: Authenticates a user via Atlassian Confluence Server JWT. Redirects to the booking UI on success. No direct authentication required (whitelisted route).
      operationId: confluenceServerLogin
      security: []
      parameters:
        - name: orgID
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Organization ID
        - name: jwt
          in: path
          required: true
          schema:
            type: string
          description: JWT signed with the Confluence shared secret
      responses:
        "307":
          description: Redirect to booking UI (success or failure page)
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: Organization not found (plain text response)

  # ===========================
  # Healthcheck
  # ===========================
...